pkgconfig: Add APIs for reading .pc files
The binding gneerator currently needs to be passed a library
or list of libraries to generate bindings for.

It would be much more convenient, though, if a project could
just pass a pkgconfig name instead.

This commit doesn't add that feature yet, but it does add basic
support for pulling data from .pc files.

A subsequent commit will leverage that support.