tests: Add shared fixtures for unit tests
At the moment, this project has no functionality nor tests.

We're going to need both, of course, so as start, let's
sketch out the beginning of a framework for testing.

This commit introduces a conftest.py file for pytest with
reusable fixtures for temporary files and directories. These
helpers will be leveraged to provide a consistent
foundation for upcoming test cases.