project: Update commit message guidelines
The dwarfbind project has a fairly opinionated take on commit messages.
This is true for output from human contributors but also for AI ones.

The problem is the guidelines don't quite make LLMs produce the right
style of output.

This commit nails the spec down more precisely to alleviate that
problem.